voice/django-voice/requirements.txt
django
africastalking
python-dotenv